fix
add mysh_def.h
