+++ include/mysh_parser.h
@@ -1,32 +1,5 @@
 #ifndef MYSH_PARSER_H
 #define MYSH_PARSER_H
-#define DO_PREFIX 0
-#define DO_SUFFIX 1
-typedef enum {
-	PIPELINE,
-	SIMPLE_COMMAND,
-	CMD_PREFIX_HEAD,
-	CMD_PREFIX,
-	CMD_SUFFIX,
-	CMD_SUFFIX_HEAD,
-	CMD_WORD,
-	CMD_WORD_HEAD,
-	ARG_WORD,
-	IO_REDIRECT,
-	IO_NUMBER,
-	IO_FILE,
-	FILENAME,
-	IO_HERE,
-	HERE_END
-} e_node_type;
-typedef struct s_cmdlst{
-	char			*value;
-	size_t			node_num;
-	e_token_type	token_type;
-	e_node_type		node_type;
-	struct s_cmdlst	*child;
-	struct s_cmdlst	*sibling;
-} t_cmdlst;
 t_cmdlst	*parser(t_token *token);
 t_cmdlst	*cmd_suffix(t_token *token);
 t_cmdlst	*cmd_prefix(t_token *token);
