+++ src/envlst/delete_envlst_keyval.c
@@ -1,3 +1,4 @@
+#include "mysh_def.h"
 #include "mysh_envlst.h"
 #include <stdlib.h>
 #include <stdio.h>
