+++ src/envlst/lookup_envlst.c
@@ -1,4 +1,4 @@
-#include "mysh_envlst.h"
+#include "mysh_def.h"
 #include "libft.h"
 #include <stdio.h>
