+++ src/executer/executer.c
@@ -1,6 +1,5 @@
+#include "mysh_def.h"
 #include "mysh_envlst.h"
-#include "mysh_lexer.h"
-#include "mysh_parser.h"
 #include "mysh_executer.h"
 #include "mysh_pipe.h"
 #include "mysh_redirect.h"
