+++ src/executer/get_argv.c
@@ -1,9 +1,5 @@
-#include "mysh_envlst.h"
-#include "mysh_lexer.h"
-#include "mysh_parser.h"
-#include "mysh_executer.h"
+#include "mysh_def.h"
 #include "mysh_utils.h"
-#include "libft.h"
 #include <unistd.h>
 #include <stdio.h>
 #include <stdlib.h>
