+++ src/executer/get_path.c
@@ -1,7 +1,5 @@
+#include "mysh_def.h"
 #include "mysh_envlst.h"
-#include "mysh_lexer.h"
-#include "mysh_parser.h"
-#include "mysh_executer.h"
 #include "mysh_utils.h"
 #include "libft.h"
 #include <unistd.h>
