+++ src/lexer/init_lexer.c
@@ -1,4 +1,4 @@
-#include "mysh_lexer.h"
+#include "mysh_def.h"
 #include "libft.h"
 #include <stdio.h>
