+++ src/prompt/prompt.c
@@ -1,5 +1,4 @@
 #include <unistd.h>
-#include "mysh_prompt.h"
 
 void prompt(void){
 	write(1, "> ", 2);
