+++ src/utils/debug_cmdlst.c
@@ -1,5 +1,4 @@
-#include "mysh_lexer.h"
-#include "mysh_parser.h"
+#include "mysh_def.h"
 #include "mysh_utils.h"
 #include "libft.h"
 #include <stdlib.h>
